chore: modernize project to uv + pyproject.toml
Replace setup.py/tox.ini/pip-compile with pyproject.toml and uv.
- Consolidate all config (ruff, pytest, coverage, ty) into pyproject.toml
- Switch Makefile from pyenv to uv (`make check` = format+lint+typecheck+tests)
- Add ty type checker (replaces mypy), configured permissively
- Add dependency-groups for dev deps
- Mark test_controlnet as gputest (was skipif cpu, failed on MPS due to missing densepose expected output)
- Delete setup.py, tox.ini, tests/ruff.toml, tests/constraints.txt, requirements-dev.{in,txt}

+++ imaginairy/api/video_sample.py
@@ -331,7 +331,7 @@ def save_video_bounce(
     frames_pil = [Image.fromarray(frame) for frame in frames_np]
     if interpolate_fps:
         # bring it up to at least 60 fps
-        fps_multiplier = int(math.ceil(interpolate_fps / output_fps))
+        fps_multiplier = math.ceil(interpolate_fps / output_fps)
         frames_pil = interpolate_images(frames_pil, fps_multiplier=fps_multiplier)
 
     transition_duration_ms = transition_duration * 1000
