+++ imaginairy/utils/roi_utils.py
@@ -8,16 +8,16 @@ logger = logging.getLogger(__name__)
 def square_roi_coordinate(roi, max_width, max_height, best_effort=False):
     """Given a region of interest, returns a square region of interest."""
     x1, y1, x2, y2 = roi
-    x1, y1, x2, y2 = int(round(x1)), int(round(y1)), int(round(x2)), int(round(y2))
+    x1, y1, x2, y2 = round(x1), round(y1), round(x2), round(y2)
     roi_width = x2 - x1
     roi_height = y2 - y1
     if roi_width < roi_height:
         diff = roi_height - roi_width
-        x1 -= int(round(diff / 2))
+        x1 -= round(diff / 2)
         x2 += roi_height - (x2 - x1)
     elif roi_height < roi_width:
         diff = roi_width - roi_height
-        y1 -= int(round(diff / 2))
+        y1 -= round(diff / 2)
         y2 += roi_width - (y2 - y1)
 
     x1, y1, x2, y2 = move_roi_into_bounds(
@@ -48,15 +48,15 @@ def resize_roi_coordinates(
     max_expansion_factor = min(max_height / side_length_y, max_width / side_length_x)
     expansion_factor = min(expansion_factor, max_expansion_factor)
 
-    expansion_x = int(round(side_length_x * expansion_factor - side_length_x))
-    expansion_x_a = int(round(expansion_x / 2))
+    expansion_x = round(side_length_x * expansion_factor - side_length_x)
+    expansion_x_a = round(expansion_x / 2)
     expansion_x_b = expansion_x - expansion_x_a
     x1 -= expansion_x_a
     x2 += expansion_x_b
 
-    expansion_y = int(round(side_length_y * expansion_factor - side_length_y))
+    expansion_y = round(side_length_y * expansion_factor - side_length_y)
     if expand_up:
-        expansion_y_a = int(round(expansion_y / 2))
+        expansion_y_a = round(expansion_y / 2)
         expansion_y_b = expansion_y - expansion_y_a
         y1 -= expansion_y_a
         y2 += expansion_y_b
@@ -89,7 +89,7 @@ def move_roi_into_bounds(roi, max_width, max_height, force=False):
     if y2 > max_height:
         y1 -= y2 - max_height
         y2 = max_height
-    x1, y1, x2, y2 = int(round(x1)), int(round(y1)), int(round(x2)), int(round(y2))
+    x1, y1, x2, y2 = round(x1), round(y1), round(x2), round(y2)
     # Force ROI to fit within image boundaries (sacrificing size and aspect ratio of ROI)
     if force:
         x1 = max(0, x1)
