+++ tests/test_utils/test_model_cache.py
@@ -19,7 +19,7 @@ class DummyMemoryModule(nn.Module):
 def create_model_of_n_bytes(n):
     import math
 
-    n = int(math.floor(n / 4))
+    n = math.floor(n / 4)
     return DummyMemoryModule(n)
